+++ Partisans/Classes/ScoreViewController.m
@@ -98,34 +98,25 @@
     NSUInteger successCount = envoy.successfulMissionCount;
     NSUInteger failCount = envoy.failedMissionCount;
     
-    NSString *string = [[NSString alloc] initWithFormat:@"%d", successCount];
+    NSString *string = [[NSString alloc] initWithFormat:@"%lu", (unsigned long)successCount];
     self.scoreLabel1.text = string;
     
-    string = [[NSString alloc] initWithFormat:@"%d", failCount];
+    string = [[NSString alloc] initWithFormat:@"%lu", (unsigned long)failCount];
     self.scoreLabel2.text = string;
     
-    
-    if (successCount == 1)
-    {
+    if (successCount == 1) {
         self.titleLabel1.text = NSLocalizedString(@"Successful Mission", @"Successful Mission  --  label");
-    }
-    else
-    {
+    } else {
         self.titleLabel1.text = NSLocalizedString(@"Successful Missions", @"Successful Missions  --  label");
     }
     
-    if (failCount == 1)
-    {
+    if (failCount == 1) {
         self.titleLabel2.text = NSLocalizedString(@"Failed Mission", @"Failed Mission  --  label");
-    }
-    else
-    {
+    } else {
         self.titleLabel2.text = NSLocalizedString(@"Failed Missions", @"Failed Missions  --  label");
     }
     
-    
-    switch (successCount)
-    {
+    switch (successCount) {
         case 0:
             self.summaryLabel1.text = NSLocalizedString(@"We need to complete three missions to ensure the success of our movement.", @"We need to complete three missions to ensure the success of our movement.  --  summary");
             break;
@@ -146,8 +137,7 @@
             break;
     }
     
-    switch (failCount)
-    {
+    switch (failCount) {
         case 0:
             self.summaryLabel2.text = NSLocalizedString(@"The Operatives will need to sabotage three missions to prevent our success.", @"The Operatives will need to sabotage three missions to prevent our success.  --  summary");
             break;
@@ -174,8 +164,7 @@
 
 - (IBAction)setupButtonPressed:(id)sender
 {
-    if (!self.setupMenuItems)
-    {
+    if (!self.setupMenuItems) {
         SetupGameMenuItems *items = [[SetupGameMenuItems alloc] init];
         self.setupMenuItems = items;
     }
